Fix 3 bugs in SDMMC one block reading function. They are important for SD card operation. 1) Before DCTRL register is set, it must insert delay. 2) When SDMMC data transfer is operated, the hard flow control is necessary. 3) When data is read from SD card, It needn't judge the RXFIFOHF in SDMMC_STA register. HAL of ST is not good in SDMMC polling read function. Now, SD card one block R/W operator can work correctly!!!

+++ rtos/TEST_FreeRTOS/Driver/inc/stm32f746g_disco.h
@@ -104,7 +104,7 @@ void drawString(
 
 uint32_t getRandomData(void);
 bool_t checkDMA(uint16_t data);
-bool_t checkSDMMC(const uint8_t data);
+bool_t checkSDMMC(const uint16_t data);
 
 #ifdef __cplusplus
 }
